+++ src/authorize.rb
@@ -1,5 +1,7 @@
 # Configure the public key for SSH access
 class Authorize
+  HOME_PATH = '/home/vagrant'.freeze
+
   attr_accessor :config, :settings
 
   def initialize(config, settings)
@@ -8,13 +10,16 @@ class Authorize
   end
 
   def configure
-    return unless settings.include? 'authorize'
+    unless settings.include?('authorize') && File.exist?(File.expand_path(settings['authorize']))
+      return
+    end
 
-    if File.exist? File.expand_path(settings['authorize'])
-      config.vm.provision 'shell' do |s|
-        s.inline = 'echo $1 | grep -xq "$1" /home/vagrant/.ssh/authorized_keys || echo "\n$1" | tee -a /home/vagrant/.ssh/authorized_keys'
-        s.args = [File.read(File.expand_path(settings['authorize']))]
-      end
+    keys = "#{HOME_PATH}/.ssh/authorized_keys"
+
+    config.vm.provision :shell do |s|
+      s.name   = 'Configure the public key for SSH access'
+      s.inline = "echo $1 | grep -xq \"$1\" #{keys} || echo \"\n$1\" | tee -ia #{keys}"
+      s.args   = [File.read(File.expand_path(settings['authorize']))]
     end
   end
 end